de if a switch

+++ CombateElVirus/src/Fondo.h
@@ -9,7 +9,7 @@ class Fondo {
 private:
 
 public:
-	void Escenario(float nivel);
-	void Titulo(float nivel);
-	void Musica(float nivel);//No funciona
+	void Escenario(int nivel);
+	void Titulo(int nivel);
+	void Musica(int nivel);//No funciona
 };
